srt/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
Flask==3.1.0
Flask-Cors==5.0.0
future==1.0.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
MarkupSafe==3.0.2
openai==1.57.0
packaging==24.2
pydantic==2.10.3
pydantic_core==2.27.1
python-multipart==0.0.19
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
certifi
